fix: add more examples and correct some confusing wording in cli options' descriptions (#43)
As fixing -h/--help output in README file was mentioned in other PR, couple things in option descriptions did jump out to me as I skimmed over them earlier, and thought it'd be faster to propose changes rather than describe them.
Feel free to pick/commit couple ones that make sense and discard the rest, as I'm not a writer of any kind, so what do I know.

E.g. chmod option did bother me in particular:

- '0700' is a bad example - log files should not be made executable, pretty much ever, and people tend to copy bad examples without understanding, noticing or just giving it enough thought.

- "chmod the logfile to this value" -> "change file mode" since chmod isn't a common verb afaik.

- "or NULL" - raw C defaults probably shouldn't be documented like that :)

And so on, though obviously nitpicks, since imo having basically any description does 99% of the job as it is.